refactor: change TouchableOpacity to Pressable, use LinkButton on SectionHeader
