+++ components/ui/home/index/molecules/SectionHeader.tsx
@@ -1,7 +1,7 @@
-import { Text, TouchableOpacity, View } from "react-native";
-import { Link } from "expo-router";
-import HOME_MESSAGE from "@/message/home";
+import { Text, View } from "react-native";
 import type { HomeSectionItemProps } from "@/types/screen/home";
+import HOME_MESSAGE from "@/message/home";
+import LinkButton from "@/components/common/navigation/LinkButton";
 
 export default function HomeSectionHeader({ id }: HomeSectionItemProps) {
   return (
@@ -9,20 +9,15 @@ export default function HomeSectionHeader({ id }: HomeSectionItemProps) {
       <Text tw="font-bold text-xl text-gray-900">
         {HOME_MESSAGE.homeSectionTitle[id]}
       </Text>
-      <Link
+      <LinkButton
         href={
           id === "travel-log" || id === "review"
             ? `/(tabs)/(travelReview)/${id}`
             : `/(tabs)/home/${id}`
         }
-        asChild
-      >
-        <TouchableOpacity>
-          <Text tw="text-xs text-gray-300">
-            {HOME_MESSAGE.homeSectionTitle.more}
-          </Text>
-        </TouchableOpacity>
-      </Link>
+        text="더보기"
+        textStyle="font-normal text-xs text-gray-300"
+      />
     </View>
   );
 }
